Move exception api route to new file in views
